+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file.
 
+# [2.9.3] - 13-08-2019
+
+### Fixes
+
+- Remove double await typo in Payments
+
+# [2.9.2] - 13-08-2019
+
+### Fixes
+
+- Fix isAuthenticated check in Notifications getIotToken
+
+# [2.9.1] - 13-08-2019
+
+### Fixes
+
+- Await checkStatus to include error response
+
+# [2.9.0] - 12-08-2019
+
+### Added
+
+- Direct Debit methods in Payments
+
 # [2.8.3] - 09-07-2019
 
 ### Fixes